requirements.txt
discord.py
requests
python-dotenv
py-cord
pytz
sentry_sdk
gql[all]